vrvp-strategy/requirements.txt
# VRVP Forex Trading Strategy Dependencies

# Data Analysis
pandas>=2.0.0
numpy>=1.24.0

# Technical Indicators
pandas-ta>=0.3.14b0

# Smart Money Concepts (FVG detection)
smartmoneyconcepts>=0.0.21

# Volume Profile
MarketProfile>=0.2.0

# Backtesting
backtrader>=1.9.78

# OANDA API for live trading
oandapyV20>=0.6.3

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# Logging and Monitoring
loguru>=0.7.0

# Notifications (optional)
requests>=2.31.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0

# Data visualization (optional)
matplotlib>=3.7.0
